derive: move whitespace conversion into config module
